requirements.txt
fastapi
sqlalchemy
uvicorn
python-binance
Jinja2